pyproject.toml
[project]
name = "projeto-agente-ai"
version = "0.1.0"
description = "Projeto de Agente de IA utilizando o Google Gemini"
authors = [{ name = "Icaro Cesar", email = "icarosouza04@outlook.com" }]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "crewai>=1.3.0",
    "flask>=3.1.2",
    "google-genai>=1.48.0",
    "google-generativeai>=0.8.5",
    "langchain-openai>=1.0.2",
    "openai>=1.109.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "transformers>=4.57.1",
]

managed = true
